TODO: Revert back to double implementation
BigDecimal has a big disadvantage of 1) being incredibly slow
and seemingly poorly optimized by the JIT, I mean, the thing uses
strings for half of the operations, and 2) actually doing math
with BigDecimals (and initializing them) is a disaster.

I've thought about the previous implementation and realized that
the problem wasn't inherintly with the double's relatively low
precision (at least low by the scientific standards), but with
how events are dispatched with JavaFX.

I completely forgot that JavaFX either uses coroutines or multithreading
to dispatch events which causes race conditions. This led to the
graph appearing to only be zoomed to about 1e-2 while in reality
it was zoomed in much farther but the grid caught the change event
too slowly as it would appear.