REDO: Use XYChart<X, Y>
I spent a lot of time trying to implement a "proper" graph
component. There were simply too many problems to deal with that
should have been addressed earlier that hamper the current design.
Unfortunately, there is not point in trying to reimplement it when
XYChart<X,Y> will suffice for now.

I may create a separate library specifically for scientific graphing
that will be thoroughly planned and hopefully avoid many pitfalls with
the current design. For instance:

1) Handling small values outside of the origin -- how do we display them?
2) Zooming indefinitely (almost), binary64 precision isn't enough
3) BigDecimal is too slow to recompute function samples and grid guides

There may be a few more challenges, but those are the major ones that
would require an entire rewrite to amend. We don't have time for this,
so we will opt to use XYChart<X,Y> as it will serve the purpose at the
moment.

To be honest, this project served as a painful lesson that I should not
try to invent something that is unnecessary. Would it be nice to have
a bunch of features on a graph not offered by XYChart? Sure, but XYChart
will serve the purpose just fine and will infintely increase development time.

This begins the rewrite of the application using XYChart which should speed
up development and will, hopefully, result in a finished product in the
next week.